Update after rebase with Solid-Texture
- Removed "noisetexture.h" from CMakeLists.txt
- Removed noisetexture.h class file and embedded it into texture.h
- included perlin.h
